Refactor imports in test files to remove file extensions for `Artifacts`. Add new `BaseTaskArtifact`, `TextTaskArtifact`, `TableTaskArtifact`, `SpreadsheetTaskArtifact`, `ImageTaskArtifact`, and `ArtifactFactory` classes.
