+++ tests/imageManagerPhase5.test.js
@@ -30,7 +30,7 @@ class DummyBaseRequestManager {
 global.BaseRequestManager = DummyBaseRequestManager;
 
 // Import artifacts & ImageManager
-const { ArtifactFactory } = require('../src/AdminSheet/Models/Artifacts.js');
+const { ArtifactFactory } = require('../src/AdminSheet/Models/Artifacts');
 const ImageManager = require('../src/AdminSheet/RequestHandlers/ImageManager.js');
 
 function buildImageManager(){
